requirements.txt
pydantic-settings>=2.1.0
class-doc>=0.2.6